refactor(ICache): move ICacheParams into separate file
