+++ src/main/scala/xiangshan/frontend/icache/ICache.scala
@@ -47,81 +47,6 @@ import xiangshan.cache._
 import xiangshan.cache.mmu.TlbRequestIO
 import xiangshan.frontend._
 
-case class ICacheParameters(
-    nSets:               Int = 256,
-    nWays:               Int = 4,
-    rowBits:             Int = 64,
-    nTLBEntries:         Int = 32,
-    tagECC:              Option[String] = None,
-    dataECC:             Option[String] = None,
-    replacer:            Option[String] = Some("random"),
-    PortNumber:          Int = 2,
-    nFetchMshr:          Int = 4,
-    nPrefetchMshr:       Int = 10,
-    nWayLookupSize:      Int = 32,
-    DataCodeUnit:        Int = 64,
-    ICacheDataBanks:     Int = 8,
-    ICacheDataSRAMWidth: Int = 66,
-    // TODO: hard code, need delete
-    partWayNum:          Int = 4,
-    nMMIOs:              Int = 1,
-    blockBytes:          Int = 64,
-    cacheCtrlAddressOpt: Option[AddressSet] = None
-) extends L1CacheParameters {
-
-  val setBytes:     Int         = nSets * blockBytes
-  val aliasBitsOpt: Option[Int] = Option.when(setBytes > pageSize)(log2Ceil(setBytes / pageSize))
-  val reqFields: Seq[BundleFieldBase] = Seq(
-    PrefetchField(),
-    ReqSourceField()
-  ) ++ aliasBitsOpt.map(AliasField)
-  val echoFields: Seq[BundleFieldBase] = Nil
-  def tagCode:    Code                 = Code.fromString(tagECC)
-  def dataCode:   Code                 = Code.fromString(dataECC)
-  def replacement = ReplacementPolicy.fromString(replacer, nWays, nSets)
-}
-
-trait HasICacheParameters extends HasL1CacheParameters with HasInstrMMIOConst with HasIFUConst {
-  val cacheParams: ICacheParameters = icacheParameters
-
-  def ctrlUnitParamsOpt: Option[L1ICacheCtrlParams] = OptionWrapper(
-    cacheParams.cacheCtrlAddressOpt.nonEmpty,
-    L1ICacheCtrlParams(
-      address = cacheParams.cacheCtrlAddressOpt.get,
-      regWidth = XLEN
-    )
-  )
-
-  def ICacheSets:          Int = cacheParams.nSets
-  def ICacheWays:          Int = cacheParams.nWays
-  def PortNumber:          Int = cacheParams.PortNumber
-  def nFetchMshr:          Int = cacheParams.nFetchMshr
-  def nPrefetchMshr:       Int = cacheParams.nPrefetchMshr
-  def nWayLookupSize:      Int = cacheParams.nWayLookupSize
-  def DataCodeUnit:        Int = cacheParams.DataCodeUnit
-  def ICacheDataBanks:     Int = cacheParams.ICacheDataBanks
-  def ICacheDataSRAMWidth: Int = cacheParams.ICacheDataSRAMWidth
-  def partWayNum:          Int = cacheParams.partWayNum
-
-  def ICacheMetaBits:      Int = tagBits // FIXME: unportable: maybe use somemethod to get width
-  def ICacheMetaCodeBits:  Int = 1       // FIXME: unportable: maybe use cacheParams.tagCode.somemethod to get width
-  def ICacheMetaEntryBits: Int = ICacheMetaBits + ICacheMetaCodeBits
-
-  def ICacheDataBits: Int = blockBits / ICacheDataBanks
-  def ICacheDataCodeSegs: Int =
-    math.ceil(ICacheDataBits / DataCodeUnit).toInt // split data to segments for ECC checking
-  def ICacheDataCodeBits: Int =
-    ICacheDataCodeSegs * 1 // FIXME: unportable: maybe use cacheParams.dataCode.somemethod to get width
-  def ICacheDataEntryBits: Int = ICacheDataBits + ICacheDataCodeBits
-  def ICacheBankVisitNum:  Int = 32 * 8 / ICacheDataBits + 1
-  def highestIdxBit:       Int = log2Ceil(nSets) - 1
-
-  require((ICacheDataBanks >= 2) && isPow2(ICacheDataBanks))
-  require(ICacheDataSRAMWidth >= ICacheDataEntryBits)
-  require(isPow2(ICacheSets), s"nSets($ICacheSets) must be pow2")
-  require(isPow2(ICacheWays), s"nWays($ICacheWays) must be pow2")
-}
-
 class ICacheMetadata(implicit p: Parameters) extends ICacheBundle {
   val tag: UInt = UInt(tagBits.W)
 }
